backend/requirements.txt
# Python 3.10.0

#django
djangorestframework==3.15.2
django==5.1
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1

pillow==10.4.0